add tooltip to integration links for navigation clarity
This change introduces an «EuiToolTip» around the integration links in the Decoders component. The tooltip provides users with a description of the action that will occur when they click on the link, enhancing the user experience by making navigation more intuitive. This small addition aims to reduce confusion and improve accessibility within the application.